backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery & Task Queue
celery==5.3.6
flower==2.0.1
kombu==5.3.5

# AI & ML
anthropic==0.18.1
openai==1.12.0
langchain==0.1.7
langchain-community==0.0.20
sentence-transformers==2.3.1

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10

# Data validation
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP & APIs
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3

# Monitoring & Logging
structlog==24.1.0
sentry-sdk==1.40.0

# Storage
minio==7.2.3

# Development
pytest==7.4.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.2.1
mypy==1.8.0

# Environment
python-dotenv==1.0.1
